+++ app/database/IdentityDao.ts
@@ -1,8 +1,8 @@
 import { Identity } from '../utils/types';
-import { identitiesTable } from '../constants/databaseNames.json';
+import databaseNames from '../constants/databaseNames.json';
 
 export async function getAllIdentities(): Promise<Identity[]> {
-    return window.database.general.selectAll(identitiesTable);
+    return window.database.general.selectAll(databaseNames.identitiesTable);
 }
 
 export const {