+++ app/database/migrations/20210426140801_create_wallet_table.ts
@@ -1,14 +1,17 @@
 import { Knex } from 'knex';
-import { walletTable } from '~/constants/databaseNames.json';
+import databaseNames from '~/constants/databaseNames.json';
 
 export async function up(knex: Knex): Promise<void> {
-    return knex.schema.createTable(walletTable, (table: Knex.TableBuilder) => {
-        table.increments('id');
-        table.string('identifier').unique().notNullable();
-        table.string('type').notNullable();
-    });
+    return knex.schema.createTable(
+        databaseNames.walletTable,
+        (table: Knex.TableBuilder) => {
+            table.increments('id');
+            table.string('identifier').unique().notNullable();
+            table.string('type').notNullable();
+        }
+    );
 }
 
 export async function down(knex: Knex): Promise<void> {
-    return knex.schema.dropTable(walletTable);
+    return knex.schema.dropTable(databaseNames.walletTable);
 }